Rewrite of the privilege escalation
Instead of requring root to run the script, the script itself
can be invoked by a non privileged user, and the script will
invoke sudo when needed.

A local "sudo" script is added into tools and the script PATH
is appended by this directory. The user password is now expected
to reside in $HOME/.cipassword and the sudo proxy script will
invoke sudo with this password.