Support for local build
Added the "-l" option to specify that the build should be performed
from the *local files* in the Git repo of the specified path - that
is, changes need not be commited to perform a build.

Slightly changed the logic of "-r", the specification of a local repo
from which the build will be performed on a checked in version (tag,
changeset, branch). The repo is now *cloned* locally, to make sure
that rogue local files does not affect the build for CM consistency.

Signed-off-by: Stefan Berg <sfb@consultron.com>