Adaptation of the deploy phase to Fuel 6.1
The structure for dea.yaml and dha.yaml files in the config
directory has been changed so that multiple Fuel versions can
be supported.

The ci_pipeline.sh will use the Fuel major version (as found in the
ISO under version.yaml) to find the appropriate deploy config to
use. By default this will be the ones found under:

   config/<Fuel version (for example 6.1)>/default_no_ha/

The deploy target can be changed either by applying the "-a" flag
(which has a new meaning) or setting the environment variable
$DEPLOYTGT. As an example, to deploy with dea.yaml parameters valid
for an ELX installation the target "elx_no_ha" can be set and the
files used will thus be those under:

   config/<Fuel version (for example 6.1)>/elx_no_ha/

In order to easily extract the version.yaml file, a new package
dependency has been added to install.sh: p7zip-full

The instructions in install.sh has also been updated so that the
name of the libvirtd service matches that of Ubuntu 14.04, which
is "libvirt-bin" and not "libvirtd".

Signed-off-by: Stefan Berg <sfb@consultron.com>