Support for unattended installation of dependencies
In order to support automated scenarios a "-y|--yes" argument has
been added, which will disable the safety prompts of the install
script.

Another change is *not* to install Docker if it's already installed,
regrardless of version. This seems to be a less error prone than
trying to re-install a later version.

Signed-off-by: Stefan Berg <sfb@consultron.com>